cloud_run_worker/requirements.txt
pandas
fastapi
uvicorn[standard]
google-cloud-aiplatform
